+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### bugfix
 - fix e2e-tests to use the "constant" key again @TmNguyen12 [#1307](https://github.com/newrelic/nri-kubernetes/pull/1307)
 
+### enhancement
+- Add metrics for ResourceQuota
 
 ## v3.45.3 - 2025-09-29